Move SignInUp form to pages
